「画像の保存最大枚数」から「次回の画像id」へ修正しました。

+++ yubimoji_recognizer/src/create_input_image_dirs.py
@@ -16,11 +16,12 @@ def make_dir_renamed_images():
         count += 1
 
         path = renamed_images_path + '\\' + name + '_memo.txt'
+        
         if os.path.exists(path):
             pass
         else:
             f = open(path,'w', encoding = 'utf-8')
-            f.write('変更した'+ name + 'ラベルの画像データを格納しています。\n画像の保存最大枚数: 0')
+            f.write('変更した'+ name + 'ラベルの画像データを格納しています。\n次回の画像id番号: 0')
             f.close()
 
 def make_dir_new_images():
@@ -33,8 +34,12 @@ def make_dir_new_images():
         count += 1
 
         path = new_images_path + '\\' + name + '_memo.txt'
-        with open(path, 'w', encoding = 'utf-8') as f:
-            f.write('変更する前の'+ name + 'ラベルの画像データを格納しています。')
+
+        if os.path.exists(path):
+            pass
+        else:
+            with open(path, 'w', encoding = 'utf-8') as f:
+                f.write('変更する前の'+ name + 'ラベルの画像データを格納しています。')
 
 
 make_dir_renamed_images()
